Renamed AikoSexp to AikoSExpression, and refactored to make it neater and cleaner.
I made good on my threat from the last commit. SexpToken has become SExpression, and SexpArray is now SExpressionArray and is a subclass of SExpression. I reckon this will make it easier to explain in the docs.

Doing it this way also made whitespace handling simpler, and the code is a bit cleaner as a result.