Events uses Timing for accurate timing.
Timing no longer requires an explicit call to setup(). It's done
implicitly when millis() is called.

Events.loop() can be passed the current time, for testing purposes. This
obviates the need for stubbing out millis() when testing Events.

Timing's variable names are camelcased, to match the rest of the code.