Allow calling Events.removeHandler from a handler.
Previously, removing a handler from inside a handler could stuff up
the iterator being used to loop through the handlers. This is now fixed.

I've extracted the handler linked list out into the EventHandlerList
class, and added a good set of tests for it. The idea is to re-use this
class for a separate list of one-shot events.

Using Events now adds 800+ bytes to the code, but 550 bytes of this is
because we're pulling in the malloc implementation.

Events.addHandler calls malloc to allocate EventHandlers, but they won't
get freed if they're removed. On the other hand, there's currently no
way to remove a handler added this way, so it's not a big problem. :)